requirements.txt
numpy
# requirements.txt

clickhouse-driver==0.2.9
python-dotenv==1.2.1
lz4==4.4.5
clickhouse-cityhash==1.0.2.4
pandas==2.3.3
pyarrow==22.0.0
pandas-stubs==2.3.2.250926
numpy==2.3.4